requirements.txt
boto3==1.21.38
kubernetes==24.2.0
PyYAML==6.0